add loading images
