+++ routes/api/notices.js
@@ -12,9 +12,19 @@ router.get('/user', authenticate, asyncWrapper(ctrNotices.getUserNotices));
 
 router.get('/:noticeId', authenticate, asyncWrapper(ctrNotices.getNoticeById));
 
-router.post('/', authenticate, asyncWrapper(ctrNotices.addNotice));
+router.post(
+  '/',
+  authenticate,
+  upload.single('image'),
+  asyncWrapper(ctrNotices.addNotice)
+);
 
-router.put('/:noticeId', authenticate, asyncWrapper(ctrNotices.updateNotice));
+router.put(
+  '/:noticeId',
+  authenticate,
+  upload.single('image'),
+  asyncWrapper(ctrNotices.updateNotice)
+);
 
 router.patch(
   '/favorite/:noticeId',
@@ -26,13 +36,6 @@ router.delete(
   '/:noticeId',
   authenticate,
   asyncWrapper(ctrNotices.deleteNotice)
-);
-
-router.patch(
-  '/image',
-  authenticate,
-  upload.single('image'),
-  asyncWrapper(ctrNotices.uploadImage)
 );
 
 module.exports = router;
